# apps/angular-permit/src/app/models/permit.models.ts
/**
 * SPABench App A — Angular Permit-to-Work SPA
 * src/app/models/permit.models.ts
 *
 * BENCHMARK RELEVANCE:
 * ────────────────────────────────────────────────────────────────────────────
 * These TypeScript interfaces are the source of EP-A-003's parameter recovery.
 *
 * TC-P2-018 (param_ts_interface, Phase 2):
 * A tool performing source map decompilation and TypeScript interface extraction
 * can recover all 5 fields of CreatePermitRequest and associate them with the
 * POST /api/rc_permit/createPermit endpoint. This is what "parameter mining
 * from TypeScript interface definition" means in the test case specification.
 *
 * The openapi-spec.json in rest-backend lists these exact 5 fields as the
 * CreatePermitRequest schema — the source TypeScript and the API spec agree.
 *
 * EP-A-003 parameters recovered:
 *   workSiteId   (body, string, required)
 *   startDate    (body, string/date, required)
 *   endDate      (body, string/date, required)
 *   hazardLevel  (body, enum LOW|MEDIUM|HIGH, required)
 *   supervisorId (body, string, required)
 *
 * TC-P2-019 (param_generic_type_annotation, Phase 2):
 * PermitService.createPermit() is typed as:
 *   this.api.post<PermitResponse>('createPermit', request: CreatePermitRequest)
 * Tools that extract generic type arguments see `CreatePermitRequest` and can
 * look it up in this file to enumerate all fields.
 */

/**
 * Request body for POST /api/rc_permit/createPermit (EP-A-003).
 * All 5 fields are required — the REST backend validates each one and returns
 * 400 if any are missing.
 */
export interface CreatePermitRequest {
  /** Unique identifier of the work site where permit is required */
  workSiteId: string;
  /** ISO 8601 date string — permit validity start */
  startDate: string;
  /** ISO 8601 date string — permit validity end */
  endDate: string;
  /** Risk classification — drives approval routing */
  hazardLevel: 'LOW' | 'MEDIUM' | 'HIGH';
  /** Employee ID of the supervising officer */
  supervisorId: string;
}

/**
 * Response from POST /api/rc_permit/createPermit.
 * permitId follows the PTW-{uuid} format generated by the REST backend.
 */
export interface PermitResponse {
  permitId: string;
  status: PermitStatus;
  workSiteId: string;
  startDate: string;
  endDate: string;
  hazardLevel: 'LOW' | 'MEDIUM' | 'HIGH';
  supervisorId: string;
  createdAt: string;
}

/**
 * Permit status lifecycle — used in detail views and list filtering.
 */
export type PermitStatus =
  | 'DRAFT'
  | 'PENDING_APPROVAL'
  | 'APPROVED'
  | 'ACTIVE'
  | 'SUSPENDED'
  | 'CLOSED'
  | 'CANCELLED';

/**
 * Response from GET /api/rc_permit/getPermitDetails (EP-A-004).
 * Includes the full conditions array — only available via source map recovery.
 */
export interface PermitDetails extends PermitResponse {
  /** Safety conditions that must be met during the permit period */
  conditions: PermitCondition[];
  /** Audit history of status changes */
  history: PermitHistoryEntry[];
}

export interface PermitCondition {
  conditionId: string;
  description: string;
  mandatory: boolean;
  verified: boolean;
}

export interface PermitHistoryEntry {
  timestamp: string;
  fromStatus: PermitStatus;
  toStatus: PermitStatus;
  changedBy: string;
  notes?: string;
}

/**
 * Response from GET /api/rc_permit/getPermitType (EP-A-001).
 * Returns the list of permit type codes available for the current work site.
 */
export interface PermitTypeListResponse {
  permitTypes: PermitTypeCode[];
}

export type PermitTypeCode =
  | 'HOT_WORK'
  | 'CONFINED_SPACE'
  | 'ELECTRICAL_ISOLATION'
  | 'WORK_AT_HEIGHT'
  | 'EXCAVATION'
  | 'RADIATION';

/**
 * Request/response for GET /api/rc_permit/getWorkOrder (EP-A-002).
 * workOrderId is passed as a query parameter — required.
 */
export interface WorkOrderResponse {
  workOrderId: string;
  title: string;
  description: string;
  priority: 'LOW' | 'MEDIUM' | 'HIGH' | 'CRITICAL';
  assignedTo: string;
  workSiteId: string;
  estimatedHours: number;
  status: 'OPEN' | 'IN_PROGRESS' | 'COMPLETED' | 'CANCELLED';
  createdAt: string;
  updatedAt: string;
}

/**
 * Auth-related models — used by AuthService (EP-A-010).
 * FormGroup field names are recovered via TC-P2-022 (param_formgroup).
 */
export interface LoginFormValue {
  username: string;
  password: string;
  rememberMe: boolean;
}

export interface LoginResponse {
  accessToken: string;
  expiresIn: number;
  user: {
    username: string;
    role: 'operator' | 'admin' | 'supervisor';
    displayName: string;
  };
}